[ET-VK] Organize utils
Refactor our `OpUtils.*`, `StagingUtils.*`, and `Utils.*` which have accumulated helper functions during the redesign.

We split them and move them into two new folders:
1. `ops/utils`:  Used by files in `ops/` and `ops/impl/`.
2. `ops/impl/utils`:  Used only by files in `ops/impl/`.

In general, this helps to unflatten the `ops/` directory:
```
[jorgep31415@devvm15882.vll0 /data/users/jorgep31415/fbsource/xplat/executorch/backends/vulkan/runtime/graph/ops (98b0e789c)]$ tree
.
├── ExecuteNode.cpp
├── ExecuteNode.h
├── impl
│   ├── Arithmetic.cpp
│   ├── Arithmetic.h
│   ├── Staging.cpp
│   ├── Staging.h
│   └── utils
│       ├── DimUtils.h
│       ├── ScalarUtils.h
│       ├── TensorUtils.cpp
│       └── TensorUtils.h
├── OperatorRegistry.cpp
├── OperatorRegistry.h
├── PrepackNode.cpp
├── PrepackNode.h
└── utils
    ├── BindingUtils.cpp
    ├── BindingUtils.h
    ├── StagingUtils.cpp
    └── StagingUtils.h

3 directories, 18 files
```

Since we're heavily updating include declarations, we also introduce the convention to only include the VK-API via
```
#include <ATen/native/vulkan/api/api.h>
```
instead of including more specific files, e.g.,
```
#include <ATen/native/vulkan/api/Tensor.h>
```

Differential Revision: [D54690346](https://our.internmc.facebook.com/intern/diff/D54690346/)

[ghstack-poisoned]
