+++ backends/vulkan/runtime/graph/ops/OperatorRegistry.cpp
@@ -8,8 +8,6 @@
 
 #include <executorch/backends/vulkan/runtime/graph/ops/OperatorRegistry.h>
 
-#include <executorch/backends/vulkan/runtime/graph/ops/impl/Arithmetic.h>
-
 namespace at {
 namespace native {
 namespace vulkan {
