+++ backends/vulkan/runtime/graph/ops/impl/utils/DimUtils.h
@@ -12,8 +12,6 @@
 
 #include <ATen/native/vulkan/api/api.h>
 
-#include <executorch/backends/vulkan/runtime/graph/containers/Value.h>
-
 namespace at {
 namespace native {
 namespace vulkan {
@@ -73,27 +71,6 @@ uint32_t dim_at(const std::vector<int64_t>& sizes) {
 template <uint32_t N>
 uint32_t dim_at(const vTensor& v_in) {
   return dim_at<N>(v_in.sizes());
-}
-
-/*
- * For most global work group sizes, returns {4, 4, 4}, but adjusts the size for
- * 2D global work group sizes. Always maintains a total of 64 invocations
- */
-api::utils::uvec3 adaptive_work_group_size(
-    const api::utils::uvec3& global_work_group);
-
-template <typename T>
-T extract_scalar(const Value& value) {
-  if (value.isInt()) {
-    return static_cast<T>(value.toInt());
-  }
-  if (value.isDouble()) {
-    return static_cast<T>(value.toDouble());
-  }
-  if (value.isBool()) {
-    return static_cast<T>(value.toBool());
-  }
-  VK_THROW("Cannot extract scalar from Value with type ", value.type());
 }
 
 } // namespace vulkan
