+++ backends/vulkan/runtime/graph/ops/utils/BindingUtils.h
@@ -16,8 +16,6 @@ namespace at {
 namespace native {
 namespace vulkan {
 
-api::utils::ivec4 get_size_as_ivec4(const vTensor& t);
-
 void bind_tensor_to_descriptor_set(
     vTensor& tensor,
     api::PipelineBarrier& pipeline_barrier,
